+++ src/qianfan/dataset/local_data_operators/check_special_characters.py
@@ -17,10 +17,10 @@ data operator for local using
 
 from typing import Any, Dict, List, Optional, Set, Union
 
-from qianfan.dataset.local_data_operators.base_local_data_operator import (
+from qianfan.dataset.local_data_operators.base import (
     BaseLocalFilterOperator,
 )
-from qianfan.dataset.local_data_operators.local_data_operator_consts import (
+from qianfan.dataset.local_data_operators.consts import (
     _default_special_characters_set,
     _special_character_map,
 )
